youtube-video-summariser/requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
faiss-cpu
langchain_google_genai
langchain-community
youtube-transcript-api
pytube
